requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
coverage==6.4.3
Django==3.2.15
djangorestframework==3.13.1
idna==3.3
pytz==2022.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
